+++ src/com/collision/shape/Shape.java
@@ -4,30 +4,30 @@ import com.collision.Intersection;
 
 public abstract class Shape {
 
-    protected float x, y, a;
+    protected double x, y, a;
 
     protected abstract void update();
     public abstract Intersection[] getIntersections(Shape shape);
-    public abstract void translate(float x, float y);
-    public abstract void rotate(float a);
+    public abstract void translate(double x, double y);
+    public abstract void rotate(double a);
 
-    public void setPosition(float x, float y) {
+    public void setPosition(double x, double y) {
         this.x = x;
         this.y = y;
         this.update();
     }
 
-    public void setAngle(float a) {
+    public void setAngle(double a) {
         this.a = a;
         this.update();
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
         this.update();
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
         this.update();
     }
